requirements.txt
smolagents[litellm]
blender-mcp
ollama
python-dotenv
requests
fastapi
uvicorn
Pillow
